Created a "concise output" option for our batch runs. Now calculating total profits (used in concise output). Major efficiency improvement for DiscPriceKDoubleAuction.trade (note that this affects output since fewer calls are made to RNG, but no other state behavior is modified). Introduced prototype ChangingZI trader.

+++ src/TradingDemo/ZI.java
@@ -48,7 +48,7 @@ public class ZI implements TradingStyle
         if( type == AgentType.BUYER )
         {
             //return (float) (random.next() * reservationValue); 
-            return (float) (reservationValue/100) * ((myAdp.getUpperBound() 
+            return (reservationValue/100) * ((myAdp.getUpperBound() 
             - myAdp.getLowerBound())*random.next() + myAdp.getLowerBound());
         }
         else
@@ -72,7 +72,7 @@ public class ZI implements TradingStyle
         if( type == AgentType.SELLER )
         {
             //return (float) ( reservationValue + (maxSellerValue - reservationValue)*random.next() );
-            return (float) reservationValue + (reservationValue * (myAdp.getLowerBound() 
+            return reservationValue + (reservationValue * (myAdp.getLowerBound() 
             + random.next() * (myAdp.getUpperBound() - myAdp.getLowerBound() ))) /100;
         }
         else
